01-URL-Shortner/pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.1.0"
description = "A simple URL shortener"
authors = ["Astha"]
package-mode = false  # <--- This is now in the correct spot!

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.128.4"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.46"
psycopg2-binary = "^2.9.11"
python-dotenv = "^1.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"